+++ lib/src/utils/update_processing_utils.dart
@@ -6,6 +6,7 @@ import 'package:televerse/telegram.dart';
 typedef UpdateCommandHandler = Future<void> Function(MessageContext messageContext, String command);
 typedef UpdateMessageHandler = Future<void> Function(MessageContext messageContext);
 typedef UpdateStepHandler = Future<void> Function(MessageContext messageContext, StepUri stepUri);
+typedef UpdateMediaFileHandler = Future<void> Function(MessageContext messageContext, String? command);
 
 void processText(Update update, UpdateMessageHandler commandHandler) {
   final message = update.message;
@@ -16,7 +17,7 @@ void processText(Update update, UpdateMessageHandler commandHandler) {
 
 void processCommand(Update update, UpdateCommandHandler commandHandler) {
   final message = update.message;
-  final command = parseCommand(message);
+  final command = parseCommand(message?.text, message?.entities);
   if (message != null && command != null) {
     processTextInternal(message, (messageContext) => commandHandler(messageContext, command));
   }
@@ -49,16 +50,16 @@ void processCallbackQuery(Update update, UpdateStepHandler commandHandler) {
     final chatId = message.chat.id;
     final userId = chatId; // TODO: Test for groups. message.from.id returns id of opponent
     final data = callbackQuery.data;
-    final user = message.from;
+    final user = callbackQuery.user;
 
     commandHandler(
       MessageContext(
         userId: userId,
         chatId: chatId,
         original: message,
-        locale: user?.languageCode,
+        locale: user.languageCode,
         editMessageId: message.messageId,
-        username: user?.username,
+        username: user.username,
       ),
       data!, //todo
     );
@@ -109,7 +110,7 @@ void processSuccessfulPayment(Update update, UpdateMessageHandler commandHandler
   );
 }
 
-void processPhoto(Update update, UpdateMessageHandler commandHandler) {
+void processPhoto(Update update, UpdateMediaFileHandler commandHandler) {
   _processMediaFile(
     update,
     (Message? message) => message?.photo?.toMediaFiles(),
@@ -117,7 +118,7 @@ void processPhoto(Update update, UpdateMessageHandler commandHandler) {
   );
 }
 
-void processVideo(Update update, UpdateMessageHandler commandHandler) {
+void processVideo(Update update, UpdateMediaFileHandler commandHandler) {
   _processMediaFile(
     update,
     (Message? message) {
@@ -128,7 +129,7 @@ void processVideo(Update update, UpdateMessageHandler commandHandler) {
   );
 }
 
-void processSticker(Update update, UpdateMessageHandler commandHandler) {
+void processSticker(Update update, UpdateMediaFileHandler commandHandler) {
   _processMediaFile(
     update,
     (Message? message) {
@@ -142,22 +143,28 @@ void processSticker(Update update, UpdateMessageHandler commandHandler) {
 void _processMediaFile(
   Update update,
   List<MediaFile>? Function(Message?) toMediaFiles,
-  UpdateMessageHandler commandHandler,
+  UpdateMediaFileHandler commandHandler,
 ) {
   final message = update.message;
   final chatId = message?.chat.id;
   final user = message?.from;
   final userId = user?.id;
 
+  final command = parseCommand(message?.caption, message?.captionEntities);
   final files = toMediaFiles(update.message);
   if (files == null || userId == null) return;
 
-  commandHandler(MessageContext(
-    userId: userId,
-    chatId: chatId ?? userId,
-    original: message,
-    locale: user?.languageCode,
-    username: user?.username ?? '',
-    mediaFiles: files,
-  ));
+  commandHandler(
+    MessageContext(
+      userId: userId,
+      chatId: chatId ?? userId,
+      original: message,
+      text: message?.caption,
+      editMessageId: message?.messageId,
+      locale: user?.languageCode,
+      username: user?.username ?? '',
+      mediaFiles: files,
+    ),
+    command,
+  );
 }